v1.1.0betaRC1 with webui and mynode integration

+++ vpn-install.sh
@@ -13,7 +13,24 @@ if [ $# -eq 0 ] || [ "$1" = "-h" ] || [ "$1" = "-help" ]; then
   exit 1
 fi
 
-plebVPNConf="/home/admin/pleb-vpn/pleb-vpn.conf"
+# find home directory based on node implementation
+if [ -d "/mnt/hdd/mynode/pleb-vpn/" ]; then
+  homedir="/mnt/hdd/mynode/pleb-vpn"
+  execdir="/opt/mynode/pleb-vpn"
+  firewallConf="/usr/bin/mynode_firewall.sh"
+elif [ -f "/mnt/hdd/raspiblitz.conf" ]; then
+  homedir="/mnt/hdd/app-data/pleb-vpn"
+  execdir="/home/admin/pleb-vpn"
+fi
+plebVPNConf="${homedir}/pleb-vpn.conf"
+plebVPNTempConf="${homedir}/pleb-vpn.conf.tmp"
+sed '1d' $plebVPNConf > $plebVPNTempConf
+source ${plebVPNTempConf}
+sudo rm ${plebVPNTempConf}
+
+if [ "${nodetype}" = "raspiblitz" ]; then
+  source /mnt/hdd/raspiblitz.conf
+fi
 
 function setting() # FILE LINENUMBER NAME VALUE
 {
@@ -26,33 +43,40 @@ function setting() # FILE LINENUMBER NAME VALUE
   echo "# ${NAME} exists->(${settingExists})"
   if [ "${settingExists}" == "0" ]; then
     echo "# adding setting (${NAME})"
-    sudo sed -i --follow-symlinks "${LINENUMBER}i${NAME}=" ${FILE}
+    sed -i --follow-symlinks "${LINENUMBER}i${NAME}=" ${FILE}
   fi
   echo "# updating setting (${NAME}) with value(${VALUE})"
-  sudo sed -i --follow-symlinks "s/^${NAME}=.*/${NAME}=${VALUE}/g" ${FILE}
+  sed -i --follow-symlinks "s/^${NAME}=.*/${NAME}=${VALUE}/g" ${FILE}
 }
 
 status() {
-  source ${plebVPNConf}
+  local webui="${1}"
 
-  isConfig=$(ls /mnt/hdd/app-data/pleb-vpn/openvpn/ | grep -c plebvpn.conf)
+  isConfig=$(ls ${homedir}/openvpn/ | grep -c plebvpn.conf)
   message="Pleb-VPN is installed, configured, and operating as expected"
   if [ ${isConfig} -eq 0 ]; then
     isConfig="no"
   else
     isConfig="yes"
   fi
-  if [ "${plebVPN}" = "off" ]; then
-    whiptail --title "Pleb-VPN status" --msgbox "
+  if [ "${plebvpn}" = "off" ]; then
+    message="Pleb-VPN is not installed. Install Pleb-VPN by selecting Pleb-VPN from the Services Menu above."
+    if [ "${webui}" = "1" ]; then
+      echo "vpn_operating=no
+config_exists=${isConfig}
+message='${message}'" | tee ${execdir}/pleb-vpn_status.tmp
+    else
+      whiptail --title "Pleb-VPN status" --msgbox "
 Pleb-VPN installed: no
 Pleb-VPN config found: ${isConfig}
 Use menu to install Pleb-VPN.
 " 10 40
+    fi
     exit 0
   else
     currentIP=$(curl https://api.ipify.org)
-    sleep 5
-    if ! [ "${currentIP}" = "${vpnIP}" ]; then
+    sleep 1
+    if ! [ "${currentIP}" = "${vpnip}" ]; then
       vpnWorking="no"
       message="ERROR: your current IP does not match your vpnIP"
     else
@@ -77,13 +101,13 @@ Otherwise you may leak your home IP if VPN drops."
       message="ERROR: your firewall is not configured properly (must allow in/out on tun0).
 Otherwise you will not be able to send or receive with VPN on."
     fi
-    firewallallowout=$(sudo ufw status verbose | grep "${vpnIP} ${vpnPort}" | grep -c "ALLOW OUT")
+    firewallallowout=$(sudo ufw status verbose | grep "${vpnip} ${vpnport}" | grep -c "ALLOW OUT")
     if [ ${firewallallowout} -eq 1 ]; then
       firewallallowout="yes"
     else
       firewallallowout="no"
       firewallOK="no"
-      message="ERROR: your firewall is not configured properly (must allow out on ${vpnIP} ${vpnPort}/udp).
+      message="ERROR: your firewall is not configured properly (must allow out on ${vpnip} ${vpnport}/udp).
 Otherwise your VPN will not reach the server and connect."
     fi
     serviceExists=$(ls /etc/systemd/system/multi-user.target.wants/ | grep -c openvpn@plebvpn.service)
@@ -93,58 +117,70 @@ Otherwise your VPN will not reach the server and connect."
     else
       serviceExists="yes"
     fi
-    whiptail --title "Pleb-VPN status" --msgbox "
+    if [ "${webui}" = "1" ]; then
+      echo "vpn_operating=${vpnWorking}
+current_ip='${currentIP}'
+config_exists=${isConfig}
+firewall_configured=${firewallOK}
+message='${message}'" | tee ${execdir}/pleb-vpn_status.tmp
+    else
+      whiptail --title "Pleb-VPN status" --msgbox "
 VPN installed: yes
 VPN operating: ${vpnWorking}
 VPN service installed: ${serviceExists}
 VPN config file found: ${isConfig}
-VPN server IP: ${vpnIP}
-VPN server port: ${vpnPort}
+VPN server IP: ${vpnip}
+VPN server port: ${vpnport}
 Current IP (should match VPN server IP): ${currentIP}
 Firewall configuration OK: ${firewallOK}
 ${message}
 " 16 100
+    fi
     exit 0
   fi
 }
 
 on() {
   # install and configure openvpn
-  source ${plebVPNConf}
-
-  # check for openvpn.conf file
   local keepconfig="${1}"
   local isRestore="${1}"
-  isconfig=$(sudo ls /mnt/hdd/app-data/pleb-vpn/openvpn/ | grep -c plebvpn.conf)
+  local webui="${2}"
+  isconfig=$(sudo ls ${homedir}/openvpn/ | grep -c plebvpn.conf)
   if ! [ ${isconfig} -eq 0 ]; then
     if [ -z "${keepconfig}" ]; then
-      whiptail --title "Use Existing Configuration?" \
-      --yes-button "Use Existing Config" \
-      --no-button "Create New Config" \
-      --yesno "There's an existing configuration found from a previous install of openvpn. Do you wish to reuse this config file?" 10 80
-      if [ $? -eq 1 ]; then
-        keepconfig="0"
-      else
-        keepconfig="1"
+      if [ ! "{webui}" = "1" ]; then
+        whiptail --title "Use Existing Configuration?" \
+        --yes-button "Use Existing Config" \
+        --no-button "Create New Config" \
+        --yesno "There's an existing configuration found from a previous install of openvpn. Do you wish to reuse this config file?" 10 80
+        if [ $? -eq 1 ]; then
+          keepconfig="0"
+        else
+          keepconfig="1"
+        fi
       fi
     fi
   else
     keepconfig="0"    
   fi
+
   # install openvpn
   sudo apt-get -y install openvpn
+
+  # get config if not webui or selected to keep config
   if [ "${keepconfig}" = "0" ]; then
-    # remove old conf file if applicable
-    isfolder=$(sudo ls /mnt/hdd/app-data/pleb-vpn/ | grep -c openvpn)
-    if ! [ ${isfolder} -eq 0 ]; then
-      sudo rm -rf /mnt/hdd/app-data/pleb-vpn/openvpn
-    fi
-    # get new conf file
-    # get local ip
-    source <(/home/admin/config.scripts/internet.sh status local)
-    # upload plebvpn.conf
-    filename=""
-    while [ "${filename}" = "" ]
+    if [ ! "{webui}" = "1" ]; then
+      # remove old conf file if applicable
+      isfolder=$(sudo ls ${homedir}/ | grep -c openvpn)
+      if ! [ ${isfolder} -eq 0 ]; then
+        sudo rm -rf ${homedir}/openvpn
+      fi
+      # get new conf file
+      # get local ip
+      localip=$(hostname -I | awk '{print $1}')
+      # upload plebvpn.conf
+      filename=""
+      while [ "${filename}" = "" ]
       do
         clear
         echo "********************************"
@@ -183,18 +219,20 @@ on() {
           echo "# USER CANCEL"
           exit 1
         fi
-    done
-    # move plebvpn.conf
-    sudo mkdir /mnt/hdd/app-data/pleb-vpn/openvpn
-    sudo mv /home/admin/plebvpn.conf /mnt/hdd/app-data/pleb-vpn/openvpn/plebvpn.conf
+      done
+      # move plebvpn.conf
+      sudo mkdir ${homedir}/openvpn
+      sudo mv /home/admin/plebvpn.conf ${homedir}/openvpn/plebvpn.conf
+    fi
   fi
+
   # get vpnIP for pleb-vpn.conf
-  vpnIP=$(cat /mnt/hdd/app-data/pleb-vpn/openvpn/plebvpn.conf | grep remote | sed 's/remote-.*$//g' | cut -d " " -f2)
-  vpnPort=$(cat /mnt/hdd/app-data/pleb-vpn/openvpn/plebvpn.conf | grep remote | sed 's/remote-.*$//g' | cut -d " " -f3)
-  setting ${plebVPNConf} "2" "vpnPort" "'${vpnPort}'"
-  setting ${plebVPNConf} "2" "vpnIP" "'${vpnIP}'"
+  vpnip=$(cat ${homedir}/openvpn/plebvpn.conf | grep remote | sed 's/remote-.*$//g' | cut -d " " -f2)
+  vpnport=$(cat ${homedir}/openvpn/plebvpn.conf | grep remote | sed 's/remote-.*$//g' | cut -d " " -f3)
+  setting ${plebVPNConf} "2" "vpnport" "'${vpnport}'"
+  setting ${plebVPNConf} "2" "vpnip" "'${vpnip}'"
   # copy plebvpn.conf to /etc/openvpn
-  sudo cp -p /mnt/hdd/app-data/pleb-vpn/openvpn/plebvpn.conf /etc/openvpn/
+  sudo cp -p ${homedir}/openvpn/plebvpn.conf /etc/openvpn/
   # fix permissions
   sudo chown admin:admin /etc/openvpn/plebvpn.conf
   sudo chmod 644 /etc/openvpn/plebvpn.conf
@@ -205,15 +243,17 @@ on() {
   sleep 10
   currentIP=$(curl https://api.ipify.org)
   sleep 10
-  if ! [ "${currentIP}" = "${vpnIP}" ]; then
+  if ! [ "${currentIP}" = "${vpnip}" ]; then
     echo "error: vpn not working"
     echo "your current IP is not your vpn IP"
     exit 1
   else
     echo "OK ... your vpn is now active"
-  fi 
+  fi
+
   # configure firewall
   echo "configuring firewall"
+  LAN=$(ip rou | grep default | cut -d " " -f3 | sed 's/^\(.*\)\.\(.*\)\.\(.*\)\.\(.*\)$/\1\.\2\.\3/g')
   # disable firewall 
   sudo ufw disable
   # allow local lan ssh
@@ -223,58 +263,72 @@ on() {
   sudo ufw default deny outgoing
   sudo ufw default deny incoming
   # allow out on openvpn
-  sudo ufw allow out to ${vpnIP} port ${vpnPort} proto udp
+  sudo ufw allow out to ${vpnip} port ${vpnport} proto udp
   # force traffic to use openvpn
   sudo ufw allow out on tun0 from any to any
   sudo ufw allow in on tun0 from any to any
   # enable firewall
   sudo ufw --force enable
-  # check firewall
-  # skip test if restore
-  if [ ! "${isRestore}" = "1" ]; then
-    echo "checking configuration"
-    echo "stop vpn"
-    sudo systemctl stop openvpn@plebvpn
-    echo "vpn stopped"
-    echo "checking firewall"
-    currentIP=$(curl https://api.ipify.org)
-    echo "current IP = (${currentIP})...should be blank"
-    if [ "${currentIP}" = "" ]; then
-      echo "firewall config ok"
-    else 
-      echo "error...firewall not configured. Clearnet accessible when VPN is off. Uninstall and re-install pleb-vpn"
-      sudo systemctl start openvpn@plebvpn
-      exit 1
-    fi
-    echo "start vpn"
-    sudo systemctl start openvpn@plebvpn
-    sleep 10
-    currentIP=$(curl https://api.ipify.org)
-    if ! [ "${currentIP}" = "${vpnIP}" ]; then
-      echo "error: vpn not working"
-      echo "your current IP is not your vpn IP"
-      exit 1
-    else
-      echo "OK ... your vpn is now active"
-    fi 
+  if [ "${nodetype}" = "mynode" ]; then
+    # add new rules to firewallConf
+    sectionLine=$(cat ${firewallConf} | grep -n "^\# Add firewall rules" | cut -d ":" -f1 | head -n 1)
+    insertLine=$(expr $sectionLine + 1)
+    sed -i "${insertLine}iufw allow in to ${LAN}.0/24" ${firewallConf}
+    sed -i "${insertLine}iufw allow out to ${LAN}.0/24" ${firewallConf}
+    sed -i "${insertLine}iufw allow out to ${vpnip} port ${vpnport} proto udp" ${firewallConf}
+    sed -i "${insertLine}iufw allow out on tun0 from any to any" ${firewallConf}
+    sed -i "${insertLine}iufw allow in on tun0 from any to any" ${firewallConf}
+    sed -i "s/ufw default allow outgoing/ufw default deny outgoing/g" ${firewallConf}
   fi
-  setting ${plebVPNConf} "2" "plebVPN" "on"
+  # check firewall
+  # skip test if restore or webui
+  if [ ! "${isRestore}" = "1" ]; then
+    if [ ! "${webui}" = "1" ]; then
+      echo "checking configuration"
+      echo "stop vpn"
+      sudo systemctl stop openvpn@plebvpn
+      echo "vpn stopped"
+      echo "checking firewall"
+      currentIP=$(curl https://api.ipify.org)
+      echo "current IP = (${currentIP})...should be blank"
+      if [ "${currentIP}" = "" ]; then
+        echo "firewall config ok"
+      else 
+        echo "error...firewall not configured. Clearnet accessible when VPN is off. Uninstall and re-install pleb-vpn"
+        sudo systemctl start openvpn@plebvpn
+        exit 1
+      fi
+      echo "start vpn"
+      sudo systemctl start openvpn@plebvpn
+      sleep 10
+      currentIP=$(curl https://api.ipify.org)
+      if ! [ "${currentIP}" = "${vpnIP}" ]; then
+        echo "error: vpn not working"
+        echo "your current IP is not your vpn IP"
+        exit 1
+      else
+        echo "OK ... your vpn is now active"
+      fi
+    fi
+  fi
+  setting ${plebVPNConf} "2" "plebvpn" "on"
   echo "OK ... plebvpn installed and configured!"
   exit 0
 }
 
 off() {
   # remove and uninstall openvpn
-  source ${plebVPNConf}
+  local webui="${1}"
 
   # first ensure that no nodes are operating on clearnet and wireguard and letsencrypt are uninstalled
-  if [ "${lndHybrid}" = "on" ] || [ "${clnHybrid}" = "on" ] || [ "${wireguard}" = "on" ] || [ "${letsencrypt_ssl}" = "on" ]; then
+  if [ "${lndhybrid}" = "on" ] || [ "${clnhybrid}" = "on" ] || [ "${wireguard}" = "on" ] || [ "${letsencrypt_ssl}" = "on" ] || [ "${torsplittunnel}" = "on" ]; then
     echo "# WARNING #"
     echo "you must first disable hybrid mode on your node(s) before removing openvpn"
     echo "otherwise your home IP will be visible"
     echo "you must also disable wireguard and letsencrypt, as they will not function without a static ip"
     exit 1
   fi
+
   # uninstall openvpn
   sudo apt-get -y purge openvpn 
   sudo rm -rf /etc/openvpn
@@ -286,21 +340,46 @@ off() {
   sudo ufw default allow outgoing
   sudo ufw default deny incoming
   # remove openvpn rule
-  sudo ufw delete allow out to ${vpnIP} port ${vpnPort} proto udp
+  sudo ufw delete allow out to ${vpnip} port ${vpnport} proto udp
   # delete force traffic to use openvpn
   sudo ufw delete allow out on tun0 from any to any
   sudo ufw delete allow in on tun0 from any to any
   # enable firewall
   sudo ufw --force enable
-  setting ${plebVPNConf} "2" "vpnPort" "''"
-  setting ${plebVPNConf} "2" "vpnIP" "''"
-  setting ${plebVPNConf} "2" "plebVPN" "off"
+  if [ "${nodetype}" = "mynode" ]; then
+    # delete rules from firewallConf
+    LAN=$(ip rou | grep default | cut -d " " -f3 | sed 's/^\(.*\)\.\(.*\)\.\(.*\)\.\(.*\)$/\1\.\2\.\3/g')
+    while [ $(cat ${firewallConf} | grep -c "ufw allow in to ${LAN}") -gt 0 ];
+    do
+      sed -i "/ufw allow in to ${LAN}\.0\/24/d" ${firewallConf}
+    done
+    while [ $(cat ${firewallConf} | grep -c "ufw allow out to ${LAN}") -gt 0 ];
+    do
+      sed -i "/ufw allow out to ${LAN}\.0\/24/d" ${firewallConf}
+    done
+    while [ $(cat ${firewallConf} | grep -c "ufw allow out to ${vpnip} port ${vpnport} proto udp") -gt 0 ];
+    do
+      sed -i "/ufw allow out to ${vpnip} port ${vpnport} proto udp/d" ${firewallConf}
+    done
+    while [ $(cat ${firewallConf} | grep -c "ufw allow out on tun0 from any to any") -gt 0 ];
+    do
+      sed -i "/ufw allow out on tun0 from any to any/d" ${firewallConf}
+    done
+    while [ $(cat ${firewallConf} | grep -c "ufw allow in on tun0 from any to any") -gt 0 ];
+    do
+      sed -i "/ufw allow in on tun0 from any to any/d" ${firewallConf}
+    done
+    sed -i "s/ufw default deny outgoing/ufw default allow outgoing/g" ${firewallConf}
+  fi
+  setting ${plebVPNConf} "2" "vpnport" "''"
+  setting ${plebVPNConf} "2" "vpnip" "''"
+  setting ${plebVPNConf} "2" "plebvpn" "off"
   exit 0
 }
 
 case "${1}" in
-  status) status ;;
-  on) on "${2}" ;;
-  off) off ;;
+  status) status "${2}" ;;
+  on) on "${2}" "${3}" ;;
+  off) off "${2}" ;;
   *) echo "err=Unknown action: ${1}" ; exit 1 ;;
 esac
